+++ index.html
@@ -13,7 +13,7 @@
   <!--[if lt IE 9]>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.3/html5shiv.js"></script>
   <![endif]-->
-  <script type="module" src="./js/index.js" async></script>
+  <script src="js/index.js" async></script>
 </head>
 
 <body>
